+++ MParT/TrainMapAdaptive.h
